Transformation: Re-organise SCC components to avoid cycles
