+++ transformations/transformations/__init__.py
@@ -10,9 +10,12 @@ from importlib.metadata import version, PackageNotFoundError
 from transformations.derived_types import * # noqa
 from transformations.argument_shape import * # noqa
 from transformations.data_offload import * # noqa
+from transformations.single_column_annotate import * # noqa
+from transformations.single_column_base import * # noqa
 from transformations.single_column_claw import * # noqa
 from transformations.single_column_coalesced import * # noqa
 from transformations.single_column_coalesced_vector import * # noqa
+from transformations.single_column_hoist import * # noqa
 from transformations.utility_routines import * # noqa
 from transformations.scc_cuf import * # noqa
 from transformations.pool_allocator import * # noqa
