+++ transformations/transformations/single_column_annotate.py
@@ -1,0 +1,339 @@
+# (C) Copyright 2018- ECMWF.
+# This software is licensed under the terms of the Apache Licence Version 2.0
+# which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
+# In applying this licence, ECMWF does not waive the privileges and immunities
+# granted to it by virtue of its status as an intergovernmental organisation
+# nor does it submit to any jurisdiction.
+
+import re
+
+from loki.expression import symbols as sym
+from loki import (
+    Transformation, FindNodes, Transformer, info, pragmas_attached,
+    as_tuple, flatten, ir, DerivedType, FindVariables,
+    CaseInsensitiveDict, pragma_regions_attached, PragmaRegion,
+    is_loki_pragma
+)
+from transformations.single_column_base import SCCBaseTransformation
+
+
+__all__ = ['SCCAnnotateTransformation']
+
+
+class SCCAnnotateTransformation(Transformation):
+    """
+    A set of utilities to insert offload directives. This includes both :any:`Loop` and
+    :any:`Subroutine` level annotations.
+
+    Parameters
+    ----------
+    horizontal : :any:`Dimension`
+        :any:`Dimension` object describing the variable conventions used in code
+        to define the horizontal data dimension and iteration space.
+    vertical : :any:`Dimension`
+        :any:`Dimension` object describing the variable conventions used in code
+        to define the vertical dimension, as needed to decide array privatization.
+    block_dim : :any:`Dimension`
+        Optional ``Dimension`` object to define the blocking dimension
+        to use for hoisted column arrays if hoisting is enabled.
+    directive : string or None
+        Directives flavour to use for parallelism annotations; either
+        ``'openacc'`` or ``None``.
+    """
+
+    def __init__(self, horizontal, vertical, directive, block_dim):
+        self.horizontal = horizontal
+        self.vertical = vertical
+        self.directive = directive
+        self.block_dim = block_dim
+
+    @classmethod
+    def kernel_annotate_vector_loops_openacc(cls, routine, horizontal, vertical):
+        """
+        Insert ``!$acc loop vector`` annotations around horizontal vector
+        loops, including the necessary private variable declarations.
+
+        Parameters
+        ----------
+        routine : :any:`Subroutine`
+            The subroutine in the vector loops should be removed.
+        horizontal: :any:`Dimension`
+            The dimension object specifying the horizontal vector dimension
+        vertical: :any:`Dimension`
+            The dimension object specifying the vertical loop dimension
+        """
+
+        # Find any local arrays that need explicitly privatization
+        argument_map = CaseInsensitiveDict({a.name: a for a in routine.arguments})
+        private_arrays = [v for v in routine.variables if not v.name in argument_map]
+        private_arrays = [v for v in private_arrays if isinstance(v, sym.Array)]
+        private_arrays = [v for v in private_arrays if not any(vertical.size in d for d in v.shape)]
+        private_arrays = [v for v in private_arrays if not any(horizontal.size in d for d in v.shape)]
+
+        if private_arrays:
+            # Log private arrays in vector regions, as these can impact performance
+            info(
+                f'[Loki-SCC::Annotate] Marking private arrays in {routine.name}: '
+                f'{[a.name for a in private_arrays]}'
+            )
+
+        mapper = {}
+        with pragma_regions_attached(routine):
+            for region in FindNodes(PragmaRegion).visit(routine.body):
+                if is_loki_pragma(region.pragma, starts_with='vector-reduction'):
+                    if (reduction_clause := re.search(r'reduction\([\w:0-9 \t]+\)', region.pragma.content)):
+
+                        loops = FindNodes(ir.Loop).visit(region)
+                        assert len(loops) == 1
+                        pragma = ir.Pragma(keyword='acc', content=f'loop vector {reduction_clause[0]}')
+                        mapper[loops[0]] = loops[0].clone(pragma=(pragma,))
+                        mapper[region.pragma] = None
+                        mapper[region.pragma_post] = None
+
+        with pragmas_attached(routine, ir.Loop):
+            for loop in FindNodes(ir.Loop).visit(routine.body):
+                if loop.variable == horizontal.index and not loop in mapper:
+                    # Construct pragma and wrap entire body in vector loop
+                    private_arrs = ', '.join(v.name for v in private_arrays)
+                    pragma = ()
+                    private_clause = '' if not private_arrays else f' private({private_arrs})'
+                    pragma = ir.Pragma(keyword='acc', content=f'loop vector{private_clause}')
+                    mapper[loop] = loop.clone(pragma=(pragma,))
+
+            routine.body = Transformer(mapper).visit(routine.body)
+
+    @classmethod
+    def kernel_annotate_sequential_loops_openacc(cls, routine, horizontal, block_dim=None, ignore=()):
+        """
+        Insert ``!$acc loop seq`` annotations around all loops that
+        are not horizontal vector loops.
+
+        Parameters
+        ----------
+        routine : :any:`Subroutine`
+            The subroutine in which to annotate sequential loops
+        horizontal: :any:`Dimension`
+            The dimension object specifying the horizontal vector dimension
+        block_dim: :any: `Dimension`
+            The dimension object specifying the blocking dimension
+        ignore: list or tuple
+            Loops to be ignored for annotation
+        """
+        block_dim_index = None if block_dim is None else block_dim.index
+        with pragmas_attached(routine, ir.Loop):
+
+            for loop in FindNodes(ir.Loop).visit(routine.body):
+                # Skip loops explicitly marked with `!$loki/claw nodep`
+                if loop.pragma and any('nodep' in p.content.lower() for p in as_tuple(loop.pragma)):
+                    continue
+
+                if loop.variable != horizontal.index and loop.variable != block_dim_index and loop not in ignore:
+                    # Perform pragma addition in place to avoid nested loop replacements
+                    loop._update(pragma=(ir.Pragma(keyword='acc', content='loop seq'),))
+
+                # Warn if we detect vector insisde sequential loop nesting
+                nested_loops = FindNodes(ir.Loop).visit(loop.body)
+                loop_pragmas = flatten(as_tuple(l.pragma) for l in as_tuple(nested_loops))
+                if any('loop vector' in pragma.content for pragma in loop_pragmas):
+                    info(f'[Loki-SCC::Annotate] Detected vector loop in sequential loop in {routine.name}')
+
+    @classmethod
+    def kernel_annotate_subroutine_present_openacc(cls, routine):
+        """
+        Insert ``!$acc data present`` annotations around the body of a subroutine.
+
+        Parameters
+        ----------
+        routine : :any:`Subroutine`
+            The subroutine to which annotations will be added
+        """
+
+        # Get the names of all array and derived type arguments
+        args = [a for a in routine.arguments if isinstance(a, sym.Array)]
+        args += [a for a in routine.arguments if isinstance(a.type.dtype, DerivedType)]
+        argnames = [str(a.name) for a in args]
+
+        routine.body.prepend(ir.Pragma(keyword='acc', content=f'data present({", ".join(argnames)})'))
+        # Add comment to prevent false-attachment in case it is preceded by an "END DO" statement
+        routine.body.append((ir.Comment(text=''), ir.Pragma(keyword='acc', content='end data')))
+
+    @classmethod
+    def insert_annotations(cls, routine, horizontal, vertical):
+
+        # Mark all parallel vector loops as `!$acc loop vector`
+        cls.kernel_annotate_vector_loops_openacc(routine, horizontal, vertical)
+
+        # Mark all non-parallel loops as `!$acc loop seq`
+        cls.kernel_annotate_sequential_loops_openacc(routine, horizontal)
+
+        # Wrap the routine body in `!$acc data present` markers
+        # to ensure device-resident data is used for array and struct arguments.
+        cls.kernel_annotate_subroutine_present_openacc(routine)
+
+        # Mark routine as `!$acc routine vector` to make it device-callable
+        routine.spec.append(ir.Pragma(keyword='acc', content='routine vector'))
+
+    def transform_subroutine(self, routine, **kwargs):
+        """
+        Apply SCCAnnotate utilities to a :any:`Subroutine`.
+
+        Parameters
+        ----------
+        routine : :any:`Subroutine`
+            Subroutine to apply this transformation to.
+        role : string
+            Role of the subroutine in the call tree; should be ``"kernel"``
+        """
+
+        role = kwargs['role']
+        targets = as_tuple(kwargs.get('targets'))
+
+        if role == 'kernel':
+            self.process_kernel(routine)
+        if role == 'driver':
+            self.process_driver(routine, targets=targets)
+
+    def process_kernel(self, routine):
+        """
+        Applies the SCCAnnotate utilities to a "kernel". This consists of inserting the relevant
+        ``'openacc'`` annotations at the :any:`Loop` and :any:`Subroutine` level.
+
+        Parameters
+        ----------
+        routine : :any:`Subroutine`
+            Subroutine to apply this transformation to.
+        """
+
+        # Bail if routine is marked as sequential
+        if SCCBaseTransformation.check_routine_pragmas(routine, self.directive):
+            return
+
+        if self.directive == 'openacc':
+            self.insert_annotations(routine, self.horizontal, self.vertical)
+
+        # Remove the vector section wrappers
+        # These have been inserted by SCCDevectorTransformation
+        section_mapper = {s: s.body for s in FindNodes(ir.Section).visit(routine.body) if s.label == 'vector_section'}
+        if section_mapper:
+            routine.body = Transformer(section_mapper).visit(routine.body)
+
+    def process_driver(self, routine, targets=None):
+        """
+        Apply the relevant ``'openacc'`` annotations to the driver loop.
+
+        Parameters
+        ----------
+        routine : :any:`Subroutine`
+            Subroutine to apply this transformation to.
+        targets : list or string
+            List of subroutines that are to be considered as part of
+            the transformation call tree.
+        """
+
+        # For the thread block size, find the horizontal size variable that is available in
+        # the driver
+        num_threads = None
+        symbol_map = routine.symbol_map
+        for size_expr in self.horizontal.size_expressions:
+            if size_expr in symbol_map:
+                num_threads = size_expr
+                break
+
+        with pragmas_attached(routine, ir.Loop, attach_pragma_post=True):
+            driver_loops = SCCBaseTransformation.find_driver_loops(routine=routine, targets=targets)
+            for loop in driver_loops:
+                loops = FindNodes(ir.Loop).visit(loop.body)
+                kernel_loops = [l for l in loops if l.variable == self.horizontal.index]
+                if kernel_loops:
+                    assert not loop == kernel_loops[0]
+                self.annotate_driver(
+                    self.directive, loop, kernel_loops, self.block_dim, num_threads
+                )
+
+            if self.directive == 'openacc':
+                # Mark all non-parallel loops as `!$acc loop seq`
+                self.kernel_annotate_sequential_loops_openacc(routine, self.horizontal, self.block_dim,
+                                                              ignore=driver_loops)
+
+        # Remove the vector section wrappers
+        # These have been inserted by SCCDevectorTransformation
+        section_mapper = {s: s.body for s in FindNodes(ir.Section).visit(routine.body) if s.label == 'vector_section'}
+        if section_mapper:
+            routine.body = Transformer(section_mapper).visit(routine.body)
+
+    @classmethod
+    def device_alloc_column_locals(cls, routine, column_locals):
+        """
+        Add explicit OpenACC statements for creating device variables for hoisted column locals.
+
+        Parameters
+        ----------
+        routine : :any:`Subroutine`
+            Subroutine to apply this transformation to.
+        column_locals : list
+            List of column locals to be hoisted to driver layer
+        """
+
+        if column_locals:
+            vnames = ', '.join(v.name for v in column_locals)
+            pragma = ir.Pragma(keyword='acc', content=f'enter data create({vnames})')
+            pragma_post = ir.Pragma(keyword='acc', content=f'exit data delete({vnames})')
+            # Add comments around standalone pragmas to avoid false attachment
+            routine.body.prepend((ir.Comment(''), pragma, ir.Comment('')))
+            routine.body.append((ir.Comment(''), pragma_post, ir.Comment('')))
+
+    @classmethod
+    def annotate_driver(cls, directive, driver_loop, kernel_loops, block_dim, num_threads):
+        """
+        Annotate driver block loop with ``'openacc'`` pragmas.
+
+        Parameters
+        ----------
+        directive : string or None
+            Directives flavour to use for parallelism annotations; either
+            ``'openacc'`` or ``None``.
+        driver_loop : :any:`Loop`
+            Driver ``Loop`` to wrap in ``'opencc'`` pragmas.
+        kernel_loops : list of :any:`Loop`
+            Vector ``Loop`` to wrap in ``'opencc'`` pragmas if hoisting is enabled.
+        block_dim : :any:`Dimension`
+            Optional ``Dimension`` object to define the blocking dimension
+            to detect hoisted temporary arrays and excempt them from marking.
+        num_threads : str
+            The size expression that determines the number of threads per thread block
+        """
+
+        # Mark driver loop as "gang parallel".
+        if directive == 'openacc':
+            arrays = FindVariables(unique=True).visit(driver_loop)
+            arrays = [v for v in arrays if isinstance(v, sym.Array)]
+            arrays = [v for v in arrays if not v.type.intent]
+            arrays = [v for v in arrays if not v.type.pointer]
+
+            # Filter out arrays that are explicitly allocated with block dimension
+            sizes = block_dim.size_expressions
+            arrays = [v for v in arrays if not any(d in sizes for d in as_tuple(v.shape))]
+            private_arrays = ', '.join(set(v.name for v in arrays))
+            private_clause = '' if not private_arrays else f' private({private_arrays})'
+            vector_length_clause = '' if not num_threads else f' vector_length({num_threads})'
+
+            # Annotate vector loops with OpenACC pragmas
+            if kernel_loops:
+                for loop in as_tuple(kernel_loops):
+                    loop._update(pragma=(ir.Pragma(keyword='acc', content='loop vector'),))
+
+            if driver_loop.pragma is None or (len(driver_loop.pragma) == 1 and
+                                              driver_loop.pragma[0].keyword.lower() == "loki" and
+                                              driver_loop.pragma[0].content.lower() == "driver-loop"):
+                p_content = f'parallel loop gang{private_clause}{vector_length_clause}'
+                driver_loop._update(pragma=(ir.Pragma(keyword='acc', content=p_content),))
+                driver_loop._update(pragma_post=(ir.Pragma(keyword='acc', content='end parallel loop'),))
+
+            # add acc parallel loop gang if the only existing pragma is acc data
+            elif len(driver_loop.pragma) == 1:
+                if (driver_loop.pragma[0].keyword == 'acc' and
+                    driver_loop.pragma[0].content.lower().lstrip().startswith('data ')):
+                    p_content = f'parallel loop gang{private_clause}{vector_length_clause}'
+                    driver_loop._update(pragma=(driver_loop.pragma[0], ir.Pragma(keyword='acc', content=p_content)))
+                    driver_loop._update(pragma_post=(ir.Pragma(keyword='acc', content='end parallel loop'),
+                                              driver_loop.pragma_post[0]))
